Force float

+++ tests/trainer/logging_/test_logger_connector.py
@@ -693,7 +693,7 @@ def test_metrics_reset(tmpdir):
 def test_result_collection_on_tensor_with_mean_reduction():
     result_collection = ResultCollection(True, torch.device("cpu"))
     product = [(True, True), (False, True), (True, False), (False, False)]
-    values = torch.arange(1, 10)
+    values = torch.arange(1, 10).float()
     batches = values * values
 
     for i, v in enumerate(values):
